ui/requirements.txt
streamlit
torch
torchvision
pillow
opencv-python-headless
albumentations
numpy
matplotlib
python-dotenv
openai
